--- eldritch/assets/MythicCards/blue/index.js
@@ -1,29 +1,0 @@
-import blue1 from './blue1.png';
-import blue2 from './blue2.png';
-import blue3 from './blue3.png';
-import blue4 from './blue4.png';
-import blue5 from './blue5.png';
-import blue6 from './blue6.png';
-import blue7 from './blue7.png';
-import blue8 from './blue8.png';
-import blue9 from './blue9.png';
-import blue10 from './blue10.png';
-import blue11 from './blue11.png';
-import blue12 from './blue12.png';
-
-const cards = {
-  blue1,
-  blue2,
-  blue3,
-  blue4,
-  blue5,
-  blue6,
-  blue7,
-  blue8,
-  blue9,
-  blue10,
-  blue11,
-  blue12,
-}
-
-export default cards
