--- eldritch/assets/MythicCards/brown/index.js
@@ -1,47 +1,0 @@
-import brown1 from './brown1.png';
-import brown2 from './brown2.png';
-import brown3 from './brown3.png';
-import brown4 from './brown4.png';
-import brown5 from './brown5.png';
-import brown6 from './brown6.png';
-import brown7 from './brown7.png';
-import brown8 from './brown8.png';
-import brown9 from './brown9.png';
-import brown10 from './brown10.png';
-import brown11 from './brown11.png';
-import brown12 from './brown12.png';
-import brown13 from './brown13.png';
-import brown14 from './brown14.png';
-import brown15 from './brown15.png';
-import brown16 from './brown16.png';
-import brown17 from './brown17.png';
-import brown18 from './brown18.png';
-import brown19 from './brown19.png';
-import brown20 from './brown20.png';
-import brown21 from './brown21.png';
-
-const cards = {
-  brown1,
-  brown2,
-  brown3,
-  brown4,
-  brown5,
-  brown6,
-  brown7,
-  brown8,
-  brown9,
-  brown10,
-  brown11,
-  brown12,
-  brown13,
-  brown14,
-  brown15,
-  brown16,
-  brown17,
-  brown18,
-  brown19,
-  brown20,
-  brown21,
-}
-
-export default cards
