--- eldritch/assets/MythicCards/green/index.js
@@ -1,41 +1,0 @@
-import green1 from './green1.png';
-import green2 from './green2.png';
-import green3 from './green3.png';
-import green4 from './green4.png';
-import green5 from './green5.png';
-import green6 from './green6.png';
-import green7 from './green7.png';
-import green8 from './green8.png';
-import green9 from './green9.png';
-import green10 from './green10.png';
-import green11 from './green11.png';
-import green12 from './green12.png';
-import green13 from './green13.png';
-import green14 from './green14.png';
-import green15 from './green15.png';
-import green16 from './green16.png';
-import green17 from './green17.png';
-import green18 from './green18.png';
-
-const cards = {
-  green1,
-  green2,
-  green3,
-  green4,
-  green5,
-  green6,
-  green7,
-  green8,
-  green9,
-  green10,
-  green11,
-  green12,
-  green13,
-  green14,
-  green15,
-  green16,
-  green17,
-  green18,
-};
-
-export default cards;
