--- eldritch/data/ancients.js
@@ -1,86 +1,0 @@
-import Ancients from '../assets/Ancients/index'
-
-const ancientsData = [
-  {
-    id: 'azathoth',
-    name: 'azathoth',
-    cardFace: Ancients.azathoth,
-    firstStage: {
-      greenCards: 1,
-      blueCards: 1,
-      brownCards: 2,
-    },
-    secondStage: {
-      greenCards: 2,
-      blueCards: 1,
-      brownCards: 3,
-    },
-    thirdStage: {
-      greenCards: 2,
-      blueCards: 0,
-      brownCards: 4,
-    },
-  },
-  {
-    id: 'cthulhu',
-    name: 'cthulhu',
-    cardFace: Ancients.cthulhu,
-    firstStage: {
-      greenCards: 0,
-      blueCards: 2,
-      brownCards: 2,
-    },
-    secondStage: {
-      greenCards: 1,
-      blueCards: 0,
-      brownCards: 3,
-    },
-    thirdStage: {
-      greenCards: 3,
-      blueCards: 0,
-      brownCards: 4,
-    },
-  },
-  {
-    id: 'iogSothoth',
-    name: 'iogSothoth',
-    cardFace: Ancients.iogSothoth,
-    firstStage: {
-      greenCards: 0,
-      blueCards: 1,
-      brownCards: 2,
-    },
-    secondStage: {
-      greenCards: 2,
-      blueCards: 1,
-      brownCards: 3,
-    },
-    thirdStage: {
-      greenCards: 3,
-      blueCards: 0,
-      brownCards: 4,
-    },
-  },
-  {
-    id: 'shubNiggurath',
-    name: 'shubNiggurath',
-    cardFace: Ancients.shubNiggurath,
-    firstStage: {
-      greenCards: 1,
-      blueCards: 1,
-      brownCards: 2,
-    },
-    secondStage: {
-      greenCards: 3,
-      blueCards: 1,
-      brownCards: 2,
-    },
-    thirdStage: {
-      greenCards: 2,
-      blueCards: 0,
-      brownCards: 4,
-    },
-  },
-]
-
-export default ancientsData
