--- eldritch/data/difficulties.js
@@ -1,16 +1,0 @@
-const difficulties = [
-  {
-    id: 'easy',
-    name: 'Низкая'
-  },
-  {
-    id: 'normal',
-    name: 'Средняя'
-  },
-  {
-    id: 'hard',
-    name: 'Высокая'
-  },
-]
-
-export default difficulties
