--- eldritch/data/mythicCards/blue/index.js
@@ -1,78 +1,0 @@
-import blueCardsAssets from '../../../assets/MythicCards/blue';
-
-const cardsData = [
-  {
-    id: 'blue1',
-    cardFace: blueCardsAssets.blue1,
-    difficulty: 'hard',
-    color:'blue'
-  },
-  {
-    id: 'blue2',
-    cardFace: blueCardsAssets.blue2,
-    difficulty: 'hard',
-    color:'blue'
-  },
-  {
-    id: 'blue3',
-    cardFace: blueCardsAssets.blue3,
-    difficulty: 'easy',
-    color:'blue'
-  },
-  {
-    id: 'blue4',
-    cardFace: blueCardsAssets.blue4,
-    difficulty: 'easy',
-    color:'blue'
-  },
-  {
-    id: 'blue5',
-    cardFace: blueCardsAssets.blue5,
-    difficulty: 'easy',
-    color:'blue'
-  },
-  {
-    id: 'blue6',
-    cardFace: blueCardsAssets.blue6,
-    difficulty: 'hard',
-    color:'blue'
-  },
-  {
-    id: 'blue7',
-    cardFace: blueCardsAssets.blue7,
-    difficulty: 'normal',
-    color:'blue'
-  },
-  {
-    id: 'blue8',
-    cardFace: blueCardsAssets.blue8,
-    difficulty: 'hard',
-    color:'blue'
-  },
-  {
-    id: 'blue9',
-    cardFace: blueCardsAssets.blue9,
-    difficulty: 'normal',
-    color:'blue'
-  },
-  {
-    id: 'blue10',
-    cardFace: blueCardsAssets.blue10,
-    difficulty: 'easy',
-    color:'blue'
-  },
-  {
-    id: 'blue11',
-    cardFace: blueCardsAssets.blue11,
-    difficulty: 'normal',
-    color:'blue'
-  },
-  {
-    id: 'blue12',
-    cardFace: blueCardsAssets.blue12,
-    difficulty: 'normal',
-    color:'blue'
-  },
-]
-
-export default cardsData
