--- eldritch/data/mythicCards/brown/index.js
@@ -1,132 +1,0 @@
-import brownCardsAssets from '../../../assets/MythicCards/brown';
-
-const cardsData = [
-  {
-    id: 'brown1',
-    cardFace: brownCardsAssets.brown1,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown2',
-    cardFace: brownCardsAssets.brown2,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown3',
-    cardFace: brownCardsAssets.brown3,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown4',
-    cardFace: brownCardsAssets.brown4,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown5',
-    cardFace: brownCardsAssets.brown5,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown6',
-    cardFace: brownCardsAssets.brown6,
-    difficulty: 'hard',
-    color:'brown'
-  },
-  {
-    id: 'brown7',
-    cardFace: brownCardsAssets.brown7,
-    difficulty: 'hard',
-    color:'brown'
-  },
-  {
-    id: 'brown8',
-    cardFace: brownCardsAssets.brown8,
-    difficulty: 'hard',
-    color:'brown'
-  },
-  {
-    id: 'brown9',
-    cardFace: brownCardsAssets.brown9,
-    difficulty: 'hard',
-    color:'brown'
-  },
-  {
-    id: 'brown10',
-    cardFace: brownCardsAssets.brown10,
-    difficulty: 'hard',
-    color:'brown'
-  },
-  {
-    id: 'brown11',
-    cardFace: brownCardsAssets.brown11,
-    difficulty: 'easy',
-    color:'brown'
-  },
-  {
-    id: 'brown12',
-    cardFace: brownCardsAssets.brown12,
-    difficulty: 'easy',
-    color:'brown'
-  },
-  {
-    id: 'brown13',
-    cardFace: brownCardsAssets.brown13,
-    difficulty: 'easy',
-    color:'brown'
-  },
-  {
-    id: 'brown14',
-    cardFace: brownCardsAssets.brown14,
-    difficulty: 'easy',
-    color:'brown'
-  },
-  {
-    id: 'brown15',
-    cardFace: brownCardsAssets.brown15,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown16',
-    cardFace: brownCardsAssets.brown16,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown17',
-    cardFace: brownCardsAssets.brown17,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown18',
-    cardFace: brownCardsAssets.brown18,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown19',
-    cardFace: brownCardsAssets.brown19,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown20',
-    cardFace: brownCardsAssets.brown20,
-    difficulty: 'normal',
-    color:'brown'
-  },
-  {
-    id: 'brown21',
-    cardFace: brownCardsAssets.brown21,
-    difficulty: 'easy',
-    color:'brown'
-  },
-]
-
-export default cardsData
