--- eldritch/data/mythicCards/green/index.js
@@ -1,114 +1,0 @@
-import greenCardsAssets from '../../../assets/MythicCards/green';
-
-const cardsData = [
-  {
-    id: 'green1',
-    cardFace: greenCardsAssets.green1,
-    difficulty: 'easy',
-    color:'green'
-  },
-  {
-    id: 'green2',
-    cardFace: greenCardsAssets.green2,
-    difficulty: 'hard',
-    color:'green'
-  },
-  {
-    id: 'green3',
-    cardFace: greenCardsAssets.green3,
-    difficulty: 'hard',
-    color:'green'
-  },
-  {
-    id: 'green4',
-    cardFace: greenCardsAssets.green4,
-    difficulty: 'hard',
-    color:'green'
-  },
-  {
-    id: 'green5',
-    cardFace: greenCardsAssets.green5,
-    difficulty: 'hard',
-    color:'green'
-  },
-  {
-    id: 'green6',
-    cardFace: greenCardsAssets.green6,
-    difficulty: 'hard',
-    color:'green'
-  },
-  {
-    id: 'green7',
-    cardFace: greenCardsAssets.green7,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green8',
-    cardFace: greenCardsAssets.green8,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green9',
-    cardFace: greenCardsAssets.green9,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green10',
-    cardFace: greenCardsAssets.green10,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green11',
-    cardFace: greenCardsAssets.green11,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green12',
-    cardFace: greenCardsAssets.green12,
-    difficulty: 'easy',
-    color:'green'
-  },
-  {
-    id: 'green13',
-    cardFace: greenCardsAssets.green13,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green14',
-    cardFace: greenCardsAssets.green14,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green15',
-    cardFace: greenCardsAssets.green15,
-    difficulty: 'normal',
-    color:'green'
-  },
-  {
-    id: 'green16',
-    cardFace: greenCardsAssets.green16,
-    difficulty: 'easy',
-    color:'green'
-  },
-  {
-    id: 'green17',
-    cardFace: greenCardsAssets.green17,
-    difficulty: 'easy',
-    color:'green'
-  },
-  {
-    id: 'green18',
-    cardFace: greenCardsAssets.green18,
-    difficulty: 'easy',
-    color:'green'
-  },
-]
-
-export default cardsData
