--- eldritch/data/mythicCards/index.js
@@ -1,9 +1,0 @@
-import brownCards from './brown';
-import blueCards from './blue';
-import greenCards from './green';
-
-export {
-  brownCards,
-  blueCards,
-  greenCards
-}
