+++ eldritch/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./style.css">
+    <script src="./script.js"></script>
+    <title>Eldritch Horror</title>
+</head>
+<body>
+    <div class="page">
+        <div class="ancients">
+            <div class="azathoth"></div>
+            <div class="cthulhu"></div>
+            <div class="iogsothoth"></div>
+            <div class="shubniggurath"></div>
+        </div>
+        <div class="level">
+            <ul class="level-list">
+                <li class="level-item">Очень легкий</li>
+                <li class="level-item">Легкий</li>
+                <li class="level-item">Средний</li>
+                <li class="level-item">Тяжелый</li>
+                <li class="level-item">Очень тяжелый</li>
+            </ul>
+        </div>
+        <div class="deck"></div>
+    </div>    
+</body>
+</html>
