+++ eldritch/script.js
@@ -1,0 +1,420 @@
+const cardsBlue = [
+    {
+        id: 'blue1',
+        cardFace: './assets/MythicCards/blue/blue1.png',
+        difficulty: 'hard',
+        color: 'blue'
+    },
+    {
+        id: 'blue2',
+        cardFace: './assets/MythicCards/blue/blue.2png',
+        difficulty: 'hard',
+        color: 'blue'
+    },
+    {
+        id: 'blue3',
+        cardFace: './assets/MythicCards/blue/blue3.png',
+        difficulty: 'easy',
+        color: 'blue'
+    },
+    {
+        id: 'blue4',
+        cardFace: './assets/MythicCards/blue/blue4.png',
+        difficulty: 'easy',
+        color: 'blue'
+    },
+    {
+        id: 'blue5',
+        cardFace: './assets/MythicCards/blue/blue5.png',
+        difficulty: 'easy',
+        color: 'blue'
+    },
+    {
+        id: 'blue6',
+        cardFace: './assets/MythicCards/blue/blue6.png',
+        difficulty: 'hard',
+        color: 'blue'
+    },
+    {
+        id: 'blue7',
+        cardFace: './assets/MythicCards/blue/blue7.png',
+        difficulty: 'normal',
+        color: 'blue'
+    },
+    {
+        id: 'blue8',
+        cardFace: './assets/MythicCards/blue/blue8.png',
+        difficulty: 'hard',
+        color: 'blue'
+    },
+    {
+        id: 'blue9',
+        cardFace: './assets/MythicCards/blue/blue9.png',
+        difficulty: 'normal',
+        color: 'blue'
+    },
+    {
+        id: 'blue10',
+        cardFace: './assets/MythicCards/blue/blue10.png',
+        difficulty: 'easy',
+        color: 'blue'
+    },
+    {
+        id: 'blue11',
+        cardFace: './assets/MythicCards/blue/blue11.png',
+        difficulty: 'normal',
+        color: 'blue'
+    },
+    {
+        id: 'blue12',
+        cardFace: './assets/MythicCards/blue/blue12.png',
+        difficulty: 'normal',
+        color: 'blue'
+    },
+]
+
+const cardsBrown = [
+    {
+        id: 'brown1',
+        cardFace: './assets/MythicCards/brown/brown1.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown2',
+        cardFace: './assets/MythicCards/brown/brown2.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown3',
+        cardFace: './assets/MythicCards/brown/brown3.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown4',
+        cardFace: './assets/MythicCards/brown/brown4.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown5',
+        cardFace: './assets/MythicCards/brown/brown5.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown6',
+        cardFace: './assets/MythicCards/brown/brown6.png',
+        difficulty: 'hard',
+        color: 'brown'
+    },
+    {
+        id: 'brown7',
+        cardFace: './assets/MythicCards/brown/brown7.png',
+        difficulty: 'hard',
+        color: 'brown'
+    },
+    {
+        id: 'brown8',
+        cardFace: './assets/MythicCards/brown/brown8.png',
+        difficulty: 'hard',
+        color: 'brown'
+    },
+    {
+        id: 'brown9',
+        cardFace: './assets/MythicCards/brown/brown9.png',
+        difficulty: 'hard',
+        color: 'brown'
+    },
+    {
+        id: 'brown10',
+        cardFace: './assets/MythicCards/brown/brown10.png',
+        difficulty: 'hard',
+        color: 'brown'
+    },
+    {
+        id: 'brown11',
+        cardFace: './assets/MythicCards/brown/brown11.png',
+        difficulty: 'easy',
+        color: 'brown'
+    },
+    {
+        id: 'brown12',
+        cardFace: './assets/MythicCards/brown/brown12.png',
+        difficulty: 'easy',
+        color: 'brown'
+    },
+    {
+        id: 'brown13',
+        cardFace: './assets/MythicCards/brown/brown13.png',
+        difficulty: 'easy',
+        color: 'brown'
+    },
+    {
+        id: 'brown14',
+        cardFace: './assets/MythicCards/brown/brown14.png',
+        difficulty: 'easy',
+        color: 'brown'
+    },
+    {
+        id: 'brown15',
+        cardFace: './assets/MythicCards/brown/brown15.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown16',
+        cardFace: './assets/MythicCards/brown/brown16.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown17',
+        cardFace: './assets/MythicCards/brown/brown17.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown18',
+        cardFace: './assets/MythicCards/brown/brown18.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown19',
+        cardFace: './assets/MythicCards/brown/brown19.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown20',
+        cardFace: './assets/MythicCards/brown/brown20.png',
+        difficulty: 'normal',
+        color: 'brown'
+    },
+    {
+        id: 'brown21',
+        cardFace: './assets/MythicCards/brown/brown21.png',
+        difficulty: 'easy',
+        color: 'brown'
+    },
+]
+
+const cardsGreen = [
+    {
+        id: 'green1',
+        cardFace: './assets/MythicCards/green/green1.png',
+        difficulty: 'easy',
+        color: 'green'
+    },
+    {
+        id: 'green2',
+        cardFace: './assets/MythicCards/green/green2.png',
+        difficulty: 'hard',
+        color: 'green'
+    },
+    {
+        id: 'green3',
+        cardFace: './assets/MythicCards/green/green3.png',
+        difficulty: 'hard',
+        color: 'green'
+    },
+    {
+        id: 'green4',
+        cardFace: './assets/MythicCards/green/green4.png',
+        difficulty: 'hard',
+        color: 'green'
+    },
+    {
+        id: 'green5',
+        cardFace: './assets/MythicCards/green/green5.png',
+        difficulty: 'hard',
+        color: 'green'
+    },
+    {
+        id: 'green6',
+        cardFace: './assets/MythicCards/green/green6.png',
+        difficulty: 'hard',
+        color: 'green'
+    },
+    {
+        id: 'green7',
+        cardFace: './assets/MythicCards/green/green7.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green8',
+        cardFace: './assets/MythicCards/green/green8.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green9',
+        cardFace: './assets/MythicCards/green/green9.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green10',
+        cardFace: './assets/MythicCards/green/green10.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green11',
+        cardFace: './assets/MythicCards/green/green11.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green12',
+        cardFace: './assets/MythicCards/green/green12.png',
+        difficulty: 'easy',
+        color: 'green'
+    },
+    {
+        id: 'green13',
+        cardFace: './assets/MythicCards/green/green13.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green14',
+        cardFace: './assets/MythicCards/green/green14.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green15',
+        cardFace: './assets/MythicCards/green/green15.png',
+        difficulty: 'normal',
+        color: 'green'
+    },
+    {
+        id: 'green16',
+        cardFace: './assets/MythicCards/green/green16.png',
+        difficulty: 'easy',
+        color: 'green'
+    },
+    {
+        id: 'green17',
+        cardFace: './assets/MythicCards/green/green17.png',
+        difficulty: 'easy',
+        color: 'green'
+    },
+    {
+        id: 'green18',
+        cardFace: './assets/MythicCards/green/green18.png',
+        difficulty: 'easy',
+        color: 'green'
+    },
+]
+
+const ancientsData = [
+    {
+        id: 'azathoth',
+        name: 'azathoth',
+        cardFace: './assets/Ancients/Azathoth.png',
+        firstStage: {
+            greenCards: 1,
+            blueCards: 1,
+            brownCards: 2,
+        },
+        secondStage: {
+            greenCards: 2,
+            blueCards: 1,
+            brownCards: 3,
+        },
+        thirdStage: {
+            greenCards: 2,
+            blueCards: 0,
+            brownCards: 4,
+        },
+    },
+    {
+        id: 'cthulhu',
+        name: 'cthulhu',
+        cardFace: './assets/Ancients/Cthulthu.png',
+        firstStage: {
+            greenCards: 0,
+            blueCards: 2,
+            brownCards: 2,
+        },
+        secondStage: {
+            greenCards: 1,
+            blueCards: 0,
+            brownCards: 3,
+        },
+        thirdStage: {
+            greenCards: 3,
+            blueCards: 0,
+            brownCards: 4,
+        },
+    },
+    {
+        id: 'iogSothoth',
+        name: 'iogSothoth',
+        cardFace: './assets/Ancients/IogSothoth.png',
+        firstStage: {
+            greenCards: 0,
+            blueCards: 1,
+            brownCards: 2,
+        },
+        secondStage: {
+            greenCards: 2,
+            blueCards: 1,
+            brownCards: 3,
+        },
+        thirdStage: {
+            greenCards: 3,
+            blueCards: 0,
+            brownCards: 4,
+        },
+    },
+    {
+        id: 'shubNiggurath',
+        name: 'shubNiggurath',
+        cardFace: './assets/Ancients/ShubNiggurath.png',
+        firstStage: {
+            greenCards: 1,
+            blueCards: 1,
+            brownCards: 2,
+        },
+        secondStage: {
+            greenCards: 3,
+            blueCards: 1,
+            brownCards: 2,
+        },
+        thirdStage: {
+            greenCards: 2,
+            blueCards: 0,
+            brownCards: 4,
+        },
+    },
+]
+
+const difficulties = [
+    {
+        id: 'very easy',
+        name: 'Очень лёгкий'
+    },
+    {
+        id: 'easy',
+        name: 'Лёгкий'
+    },
+    {
+        id: 'normal',
+        name: 'Средний'
+    },
+    {
+        id: 'hard',
+        name: 'Тяжёлый'
+    },
+    {
+        id: 'very hard',
+        name: 'Очень тяжёлый'
+    },
+]
